#6 refactor test helpers
